requirements.txt
flask==2.0.1
pillow==8.3.1
torch==1.9.0
matplotlib==3.4.3
transformers==4.9.2
flask-cors==3.0.10